Refaktor. Trekk funksjon ut i utils
